Move SortType out of the account object
