+++ feature/widget/message-list/src/main/kotlin/app/k9mail/feature/widget/message/list/MessageListConfig.kt
@@ -1,6 +1,6 @@
 package app.k9mail.feature.widget.message.list
 
-import app.k9mail.legacy.account.Account.SortType
+import app.k9mail.legacy.account.SortType
 import app.k9mail.legacy.search.LocalSearch
 
 internal data class MessageListConfig(
