+++ legacy/account/src/main/java/app/k9mail/legacy/account/Account.kt
@@ -622,16 +622,6 @@ open class Account(
         AUTO,
     }
 
-    enum class SortType(val isDefaultAscending: Boolean) {
-        SORT_DATE(false),
-        SORT_ARRIVAL(false),
-        SORT_SUBJECT(true),
-        SORT_SENDER(true),
-        SORT_UNREAD(true),
-        SORT_FLAGGED(true),
-        SORT_ATTACHMENT(true),
-    }
-
     companion object {
         /**
          * Fixed name of outbox - not actually displayed.
