+++ legacy/core/src/main/java/com/fsck/k9/preferences/AccountSettingsDescriptions.java
@@ -18,7 +18,7 @@ import app.k9mail.legacy.account.FolderMode;
 import app.k9mail.legacy.account.Account.MessageFormat;
 import app.k9mail.legacy.account.Account.QuoteStyle;
 import app.k9mail.legacy.account.Account.ShowPictures;
-import app.k9mail.legacy.account.Account.SortType;
+import app.k9mail.legacy.account.SortType;
 import app.k9mail.legacy.account.Account.SpecialFolderSelection;
 import app.k9mail.legacy.di.DI;
 import com.fsck.k9.K9;
