+++ legacy/core/src/main/java/com/fsck/k9/preferences/GeneralSettingsDescriptions.java
@@ -10,7 +10,7 @@ import java.util.TreeMap;
 import android.content.Context;
 
 import app.k9mail.feature.telemetry.api.TelemetryManager;
-import app.k9mail.legacy.account.Account.SortType;
+import app.k9mail.legacy.account.SortType;
 import app.k9mail.legacy.account.AccountDefaultsProvider;
 import app.k9mail.legacy.di.DI;
 import app.k9mail.legacy.preferences.AppTheme;
